+++ core/mixins/index.js
@@ -4,6 +4,10 @@
  *
  * Copyright © 2018 Extremely Heavy Industries Inc.
  */
-export * from './Panel';
-export * from './PanelModel';
+export * from './EventSupport';
+export * from './LayoutSupport';
+export * from './ReactiveSupport';
+export * from './StableIdSupport';
+export * from './XhIdSupport';
+