+++ mobile/appcontainer/Message.js
@@ -9,6 +9,7 @@ import {HoistComponent, elemFactory} from '@xh/hoist/core';
 import {filler} from '@xh/hoist/cmp/layout';
 import {dialog} from '@xh/hoist/mobile/cmp/dialog';
 import {button} from '@xh/hoist/mobile/cmp/button';
+import {MessageModel} from '@xh/hoist/core/appcontainer/MessageModel';
 
 /**
  * Render a modal dialog
@@ -17,6 +18,8 @@ import {button} from '@xh/hoist/mobile/cmp/button';
  */
 @HoistComponent
 class Message extends Component {
+
+    static modelClass = MessageModel;
 
     render() {
         const model = this.model,