+++ mobile/cmp/button/Button.js
@@ -7,17 +7,16 @@
 
 import {Component} from 'react';
 import PT from 'prop-types';
-import {elemFactory, HoistComponent} from '@xh/hoist/core';
+import {elemFactory, HoistComponent, LayoutSupport} from '@xh/hoist/core';
 import {button as onsenButton} from '@xh/hoist/kit/onsen';
 import {hspacer} from '@xh/hoist/cmp/layout';
 
 /**
- * Wrapper around Onsen's Button component.
- * Hoist's most basic button accepts any props supported by Onsen's Button.
- *
- * Must be provided an onClick handler.
+ * Wrapper around Onsen's Button component. Adds LayoutSupport for top-level sizing and margin/padding props.
+ * Relays all other props supported by Onsen's Button.
  */
 @HoistComponent
+@LayoutSupport
 export class Button extends Component {
 
     static propTypes = {
@@ -28,8 +27,10 @@ export class Button extends Component {
         onClick: PT.func
     };
 
+    baseClassName = 'xh-button';
+
     render() {
-        const {icon, text, modifier, active, onClick, ...rest} = this.props,
+        const {icon, text, modifier, active, onClick, style, ...rest} = this.getNonLayoutProps(),
             items = [];
 
         if (icon && text) {
@@ -41,11 +42,17 @@ export class Button extends Component {
         }
 
         return onsenButton({
-            className: this.getClassName('xh-button', active ? 'xh-button-active' : ''),
             items,
             modifier,
             onClick,
-            ...rest
+
+            style: {
+                ...style,
+                ...this.getLayoutProps()
+            },
+
+            ...rest,
+            className: this.getClassName('xh-button', active ? 'xh-button-active' : '')
         });
     }